feat: extract shared dcc-mcp-marketplace crate
Move marketplace domain types, source normalization, catalog fetch,
install/uninstall logic, installed state persistence, and integrity
verification from CLI and Gateway Admin into a shared crate consumed
by both. Eliminates the logic fork between the two existing
implementations.

Changes:
- New crate dcc-mcp-marketplace with types, source, service, error
- CLI domain/marketplace.rs re-exports from shared crate
- CLI application/marketplace.rs thin wrapper delegating to shared service
- Gateway admin marketplace handlers delegate to shared MarketplaceService
- HTTP response types preserved (frontend contract unchanged)
- CatalogEntry gains version, min_core_version, install, maintainer, icon
- icon field preserved with resolve_icon_url() for raw.githubusercontent.com
- DCC_MCP_MARKETPLACE_NO_DEFAULT_SOURCES uses truthy-only matching (1/true/yes)
- Regression tests for icon resolution and env var semantics
